✅ 🎨 Moved login related files into authentication library and updated tests accordingly
