+++ apps/app/lib/main.dart
@@ -1,5 +1,6 @@
 //import 'package:example/example.dart';
 import 'package:async_redux/async_redux.dart';
+import 'package:authentication/authentication.dart';
 import 'package:general/general.dart';
 import 'package:flutter/material.dart';
 import 'package:redux_comp/redux_comp.dart';
