+++ apps/app/pubspec.yaml
@@ -51,6 +51,9 @@ dependencies:
   tradesman:
     path: '../../libs/tradesman'
 
+  authentication:
+    path: '../../libs/authentication'
+
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
