+++ libs/authentication/.flutter-plugins
@@ -1,0 +1,14 @@
+# This is a generated file; do not edit or check into version control.
+amplify_api=C:\\flutter\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\amplify_api-0.4.5\\
+amplify_auth_cognito=C:\\flutter\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\amplify_auth_cognito-0.4.5\\
+amplify_auth_cognito_android=C:\\flutter\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\amplify_auth_cognito_android-0.4.5\\
+amplify_auth_cognito_ios=C:\\flutter\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\amplify_auth_cognito_ios-0.4.5\\
+amplify_core=C:\\flutter\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\amplify_core-0.4.5\\
+amplify_datastore=C:\\flutter\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\amplify_datastore-0.4.5\\
+amplify_flutter=C:\\flutter\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\amplify_flutter-0.4.5\\
+path_provider=C:\\flutter\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\path_provider-2.0.10\\
+path_provider_android=C:\\flutter\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\path_provider_android-2.0.14\\
+path_provider_ios=C:\\flutter\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\path_provider_ios-2.0.9\\
+path_provider_linux=C:\\flutter\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\path_provider_linux-2.1.6\\
+path_provider_macos=C:\\flutter\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\path_provider_macos-2.0.6\\
+path_provider_windows=C:\\flutter\\flutter\\.pub-cache\\hosted\\pub.dartlang.org\\path_provider_windows-2.0.6\\
