+++ libs/authentication/.flutter-plugins-dependencies
@@ -1,0 +1,1 @@
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"amplify_api","path":"C:\\\\flutter\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\amplify_api-0.4.5\\\\","dependencies":["amplify_core"]},{"name":"amplify_auth_cognito_ios","path":"C:\\\\flutter\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\amplify_auth_cognito_ios-0.4.5\\\\","dependencies":["amplify_core"]},{"name":"amplify_core","path":"C:\\\\flutter\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\amplify_core-0.4.5\\\\","dependencies":[]},{"name":"amplify_datastore","path":"C:\\\\flutter\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\amplify_datastore-0.4.5\\\\","dependencies":["amplify_core"]},{"name":"amplify_flutter","path":"C:\\\\flutter\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\amplify_flutter-0.4.5\\\\","dependencies":["amplify_core"]},{"name":"path_provider_ios","path":"C:\\\\flutter\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_ios-2.0.9\\\\","dependencies":[]}],"android":[{"name":"amplify_api","path":"C:\\\\flutter\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\amplify_api-0.4.5\\\\","dependencies":["amplify_core"]},{"name":"amplify_auth_cognito_android","path":"C:\\\\flutter\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\amplify_auth_cognito_android-0.4.5\\\\","dependencies":[]},{"name":"amplify_core","path":"C:\\\\flutter\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\amplify_core-0.4.5\\\\","dependencies":[]},{"name":"amplify_datastore","path":"C:\\\\flutter\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\amplify_datastore-0.4.5\\\\","dependencies":["amplify_core"]},{"name":"amplify_flutter","path":"C:\\\\flutter\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\amplify_flutter-0.4.5\\\\","dependencies":["amplify_core"]},{"name":"path_provider_android","path":"C:\\\\flutter\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_android-2.0.14\\\\","dependencies":[]}],"macos":[{"name":"path_provider_macos","path":"C:\\\\flutter\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_macos-2.0.6\\\\","dependencies":[]}],"linux":[{"name":"path_provider_linux","path":"C:\\\\flutter\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_linux-2.1.6\\\\","dependencies":[]}],"windows":[{"name":"path_provider_windows","path":"C:\\\\flutter\\\\flutter\\\\.pub-cache\\\\hosted\\\\pub.dartlang.org\\\\path_provider_windows-2.0.6\\\\","dependencies":[]}],"web":[]},"dependencyGraph":[{"name":"amplify_api","dependencies":["amplify_core"]},{"name":"amplify_auth_cognito","dependencies":["amplify_core","amplify_auth_cognito_android","amplify_auth_cognito_ios"]},{"name":"amplify_auth_cognito_android","dependencies":[]},{"name":"amplify_auth_cognito_ios","dependencies":["amplify_core"]},{"name":"amplify_core","dependencies":[]},{"name":"amplify_datastore","dependencies":["amplify_core"]},{"name":"amplify_flutter","dependencies":["amplify_core"]},{"name":"path_provider","dependencies":["path_provider_android","path_provider_ios","path_provider_linux","path_provider_macos","path_provider_windows"]},{"name":"path_provider_android","dependencies":[]},{"name":"path_provider_ios","dependencies":[]},{"name":"path_provider_linux","dependencies":[]},{"name":"path_provider_macos","dependencies":[]},{"name":"path_provider_windows","dependencies":[]}],"date_created":"2022-05-25 12:44:53.861197","version":"2.10.5"}
