+++ libs/authentication/lib/authentication.dart
@@ -1,7 +1,3 @@
 library authentication;
 
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
+export './pages/login.dart';
