+++ libs/general/lib/general.dart
@@ -1,4 +1,3 @@
 library general;
 
-export './login.dart';
 export './theme.dart';
